fix : printing mechanism
remove the compressed buffer, since its type is an array of char it
can't represent the compression format which is binary format.

To represent it in binary format. Avoid to hold compressed data to the
compressed buffer, write it right to the standard output stream.

use fwrite to save variable in binary format into the stream.

`fwrite(&var, sizeof var, 1, stdout)` is the way to make it.